Conflicts:
	riverDatabase.db
	src/com/river/persistence/dao/interfaces/LineDAO.java
	src/config/spring-config.xml

+++ src/com/river/business/service/implementation/RafterServiceImpl.java
@@ -21,12 +21,20 @@ public class RafterServiceImpl implements RafterService{
 
 	@Override
 	public Rafter create(Rafter toCreate) {
-		return rafterDAO.create(toCreate);
+		Rafter rafter = null;
+		if(toCreate != null){
+			rafter = rafterDAO.create(toCreate);
+		}
+		return rafter;
 	}
 
 	@Override
 	public Rafter read(Rafter toRead) {
-		return rafterDAO.read(toRead);
+		Rafter rafter = null;
+		if(toRead != null && toRead.getId() != null){
+			rafter = rafterDAO.read(toRead);
+		}
+		return rafter;
 	}
 
 	@Override
@@ -36,12 +44,20 @@ public class RafterServiceImpl implements RafterService{
 
 	@Override
 	public Rafter update(Rafter toUpdate) {
-		return rafterDAO.update(toUpdate);
+		Rafter rafter = null;
+		if(toUpdate != null && toUpdate.getId() != null){
+			rafter = rafterDAO.update(toUpdate);
+		}
+		return rafter;
 	}
 
 	@Override
 	public Rafter delete(Rafter toDelete) {
-		return rafterDAO.delete(toDelete);
+		Rafter rafter = null;
+		if(toDelete != null && toDelete.getId() != null){
+			rafter = rafterDAO.delete(toDelete);
+		}
+		return rafter;
 	}
 
 }
